// src/enums/response-code.ts
export enum ResponseCode {

  ACCESS_DENIED = 100,
  LOGIN_WRONG_PASSWORD = 101,
  PERMISSION_IMPLICIT = 103,
  TOKEN_NOT_VALID = 104,
  SESSION_TIMEOUT = 105,

  SUCCESS = 200,
  REQUEST_ACCEPTED = 201,
  REQUEST_REFUSED = 202,

  INPUT_DATA_NULL = 300,
  INPUT_DATA_WRONG_FORMAT = 301,

  DATA_UNIQUE_IMPLICIT = 400,
  DATA_NOT_FOUND = 401,
  DATA_NOT_AVAILABLE = 402,
  DATA_IMPLICIT = 403,
  DATA_CONTRAINT_VIOLATED = 404,

  SERVER_INTERNAL_ERROR = 500
}